sweep: model-in-the-loop depth/orientation sweep + ROI extraction
- inkalign sweep: runs any infer command with --layer-start/--layer-end/
  --direction over a grid of exact-size depth windows and both
  orientations, scores each prediction's text-likeness, picks the best
  window and gives a confident/weak recto-verso verdict. Command is a
  token template executed without a shell, so Windows paths survive.
- inkalign extract-roi: streams a small full-depth ROI into a local
  multiscale zarr so GPU sweeps only ever see a few hundred pixels.
  Verified against the real w025 surface volume (16 MB ROI, local
  re-profile agrees with the remote heatmap).
- score-vs-offset curve plot; end-to-end sweep test with a stub runner
  recovers the planted window (+2, forward).

+++ src/inkalign/score.py
@@ -102,6 +102,6 @@ def text_likeness(img: np.ndarray, um_per_px: float) -> TextScore:
         band = (freqs >= 1.0 / pitch_band[1]) & (freqs <= 1.0 / pitch_band[0])
         chance = band.sum() / max(len(freqs) - 1, 1)
         best = float(np.percentile(ratios, 90))  # text often occupies part of a segment
-        line_score = max(0.0, (best - chance) / (1 - chance))
+        line_score = float(max(0.0, (best - chance) / (1 - chance)))
 
     return TextScore(line_score=line_score, ink_fraction=ink_fraction, windows_scored=len(ratios))
